move truncateMiddle to utils file
