+++ src/components/TransactionsTable.tsx
@@ -26,6 +26,7 @@ import {
   renderTransactionType,
 } from "../pages/Transactions/helpers";
 import {assertNever} from "../utils";
+import { truncateMiddle } from "../pages/utils";
 
 type TransactionCellProps = {
   transaction: Types.OnChainTransaction;
@@ -78,34 +79,6 @@ function TransactionGasCell({transaction}: TransactionCellProps) {
       {renderGas(transaction.gas_used)}
     </TableCell>
   );
-}
-
-export function truncateMiddle(
-  str: any,
-  frontLen: number,
-  backLen: number,
-  truncateStr: any,
-) {
-  if (str === null) {
-    return "";
-  }
-  var strLen = str.length;
-  // Setting default values
-  frontLen = ~~frontLen; // will cast to integer
-  backLen = ~~backLen;
-  truncateStr = truncateStr || "…";
-  if (
-    (frontLen === 0 && backLen === 0) ||
-    frontLen >= strLen ||
-    backLen >= strLen ||
-    frontLen + backLen >= strLen
-  ) {
-    return str;
-  } else if (backLen === 0) {
-    return str.slice(0, frontLen) + truncateStr;
-  } else {
-    return str.slice(0, frontLen) + truncateStr + str.slice(strLen - backLen);
-  }
 }
 
 function TransactionHashCell({transaction}: TransactionCellProps) {
