+++ src/pages/utils.tsx
@@ -82,3 +82,31 @@ export async function doTransaction(
   await client.waitForTransaction(transactionRes.hash);
   return transactionRes;
 }
+
+export function truncateMiddle(
+  str: any,
+  frontLen: number,
+  backLen: number,
+  truncateStr: any,
+) {
+  if (str === null) {
+    return "";
+  }
+  var strLen = str.length;
+  // Setting default values
+  frontLen = ~~frontLen; // will cast to integer
+  backLen = ~~backLen;
+  truncateStr = truncateStr || "…";
+  if (
+    (frontLen === 0 && backLen === 0) ||
+    frontLen >= strLen ||
+    backLen >= strLen ||
+    frontLen + backLen >= strLen
+  ) {
+    return str;
+  } else if (backLen === 0) {
+    return str.slice(0, frontLen) + truncateStr;
+  } else {
+    return str.slice(0, frontLen) + truncateStr + str.slice(strLen - backLen);
+  }
+}
